Bring makefiles in line with each other

+++ Makefile
@@ -7,42 +7,52 @@
 ##########################################################################
 
 ##########################################################################
-# Default: Build iGame with standard optimizations and 00 support
+# Default: Build iGame with standard optimizations and 000 support
 ##########################################################################
 all: iGame
+
+##########################################################################
+# Set up version and date properties
+##########################################################################
+
+DATE = $(shell date --iso=date)
 
 ##########################################################################
 # Compiler settings
 ##########################################################################
 CC			= m68k-amigaos-gcc
 LINK		= m68k-amigaos-gcc
-INCLUDES	= -I/opt/amiga/m68k-amigaos/include -I/opt/amiga/m68k-amigaos/ndk-include -I/opt/amiga/m68k-amigaos/include/mui
-CFLAGS			= -c -noixemul -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68000
-CFLAGS_030	= -c -mcpu=68030 -noixemul -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68030
-CFLAGS_040	= -c -mcpu=68040 -noixemul -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68040
-CFLAGS_060	= -c -mcpu=68060 -noixemul -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68060
-CFLAGS_MOS	= -c +morphos -dontwarn=-1 -O2 -c99 -DCPU_VERS=MorphOS
-CFLAGS_AOS4	= -c +aosppc -dontwarn=-1 -O2 -c99 -DCPU_VERS=AmigaOS4
+CC_PPC		= ppc-amigaos-gcc
+LINK_PPC	= ppc-amigaos-gcc
 
-DATE = $(shell date --iso=date)
+INCLUDES	= -I$(NDK_INC) -I$(MUI38_INC)
+INCLUDES_OS4= -I$(SDK_INC) -I$(MUI50_INC)
+INCLUDES_MOS= -I$(NDK_INC) -I$(MUI50_INC)
+
+CFLAGS		= -c -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68000 -DRELEASE_DATE=$(DATE)
+CFLAGS_030	= -c -mcpu=68030 -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68030 -DRELEASE_DATE=$(DATE)
+CFLAGS_040	= -c -mcpu=68040 -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68040 -DRELEASE_DATE=$(DATE)
+CFLAGS_060	= -c -mcpu=68060 -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=68060 -DRELEASE_DATE=$(DATE)
+CFLAGS_MOS	= -c -Os -fomit-frame-pointer -std=c99 -DCPU_VERS=MorphOS -DRELEASE_DATE=$(DATE)
+CFLAGS_OS4	= -c -Os -fomit-frame-pointer -std=c99 -D__USE_INLINE__ -DCPU_VERS=AmigaOS4 -DRELEASE_DATE=$(DATE)
 
 ##########################################################################
 # Builder settings
 ##########################################################################
-#MKLIB		= join
-LIBFLAGS	= -v -lamiga -lstubs -o
-LIBFLAGS_MOS	= +morphos -lamiga -lauto -o
-LIBFLAGS_AOS4	= +aosppc -lamiga -lauto -o
+#MKLIB			= join
+LIBFLAGS		= -v -lamiga -lstubs -o
+LIBFLAGS_MOS	= -v -lamiga -lstubs -o
+LIBFLAGS_OS4	= -v -lamiga -lstubs -o
 
 ##########################################################################
 # Object files which are part of iGame
 ##########################################################################
-OBJS		= src/iGameGUI.o src/iGameMain.o src/funcs.o src/strcasestr.o src/strdup.o src/iGame_cat.o
+OBJS		= src/funcs.o src/iGameGUI.o src/iGameMain.o src/strcasestr.o src/strdup.o src/iGame_cat.o
 OBJS_030	= src/funcs_030.o src/iGameGUI_030.o src/iGameMain_030.o src/strcasestr_030.o src/strdup_030.o src/iGame_cat_030.o
 OBJS_040	= src/funcs_040.o src/iGameGUI_040.o src/iGameMain_040.o src/strcasestr_040.o src/strdup_040.o src/iGame_cat_040.o
 OBJS_060	= src/funcs_060.o src/iGameGUI_060.o src/iGameMain_060.o src/strcasestr_060.o src/strdup_060.o src/iGame_cat_060.o
 OBJS_MOS	= src/funcs_MOS.o src/iGameGUI_MOS.o src/iGameMain_MOS.o src/strcasestr_MOS.o src/strdup_MOS.o src/iGame_cat_MOS.o
-OBJS_AOS4	= src/funcs_AOS4.o src/iGameGUI_AOS4.o src/iGameMain_AOS4.o src/strcasestr_AOS4.o src/strdup_AOS4.o src/iGame_cat_AOS4.o
+OBJS_OS4	= src/funcs_OS4.o src/iGameGUI_OS4.o src/iGameMain_OS4.o src/strcasestr_OS4.o src/strdup_OS4.o src/iGame_cat_OS4.o
 
 ##########################################################################
 # Rule for building
@@ -60,10 +70,10 @@ iGame.060: $(OBJS_060)
 	$(LINK) $(OBJS_060) $(LIBFLAGS) $@
 
 iGame.MOS: $(OBJS_MOS)
-	$(LINK) $(OBJS_MOS) $(LIBFLAGS_MOS) $@
+	$(LINK_PPC) $(OBJS_MOS) $(LIBFLAGS_MOS) $@
 
-iGame.AOS4: $(OBJS_AOS4)
-	$(LINK) $(OBJS_AOS4) $(LIBFLAGS_AOS4) $@
+iGame.OS4: $(OBJS_OS4)
+	$(LINK_PPC) $(OBJS_OS4) $(LIBFLAGS_OS4) $@
 
 ##########################################################################
 # generated source files
@@ -75,7 +85,7 @@ src/iGame_cat.h: src/iGame.cd src/C_h.sd
 	cd src && flexcat iGame.cd iGame_cat.h=C_h.sd
 
 ##########################################################################
-# object files (generic 00)
+# object files (generic 000)
 ##########################################################################
 src/funcs.o: src/funcs.c src/iGame_cat.h
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/funcs.c
@@ -166,44 +176,44 @@ src/iGame_cat_060.o: src/iGame_cat.c
 ##########################################################################
 
 src/funcs_MOS.o: src/funcs.c src/iGame_cat.h
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/funcs.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/funcs.c
 
 src/iGameGUI_MOS.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/iGameGUI.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/iGameGUI.c
 
 src/iGameMain_MOS.o: src/iGameMain.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/iGameMain.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/iGameMain.c
 
 src/strcasestr_MOS.o: src/strcasestr.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/strcasestr.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/strcasestr.c
 
 src/strdup_MOS.o: src/strdup.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/strdup.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/strdup.c
 
 src/iGame_cat_MOS.o: src/iGame_cat.c
-	$(CC) $(CFLAGS_MOS) $(INCLUDES) -o $@ src/iGame_cat.c
+	$(CC_PPC) $(CFLAGS_MOS) $(INCLUDES_MOS) -o $@ src/iGame_cat.c
 
 ##########################################################################
 # object files (AOS4)
 ##########################################################################
 
-src/funcs_AOS4.o: src/funcs.c src/iGame_cat.h
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/funcs.c
+src/funcs_OS4.o: src/funcs.c src/iGame_cat.h
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/funcs.c
 
-src/iGameGUI_AOS4.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/iGameGUI.c
+src/iGameGUI_OS4.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/iGameGUI.c
 
-src/iGameMain_AOS4.o: src/iGameMain.c
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/iGameMain.c
+src/iGameMain_OS4.o: src/iGameMain.c
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/iGameMain.c
 
-src/strcasestr_AOS4.o: src/strcasestr.c
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/strcasestr.c
+src/strcasestr_OS4.o: src/strcasestr.c
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/strcasestr.c
 
-src/strdup_AOS4.o: src/strdup.c
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/strdup.c
+src/strdup_OS4.o: src/strdup.c
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/strdup.c
 
-src/iGame_cat_AOS4.o: src/iGame_cat.c
-	$(CC) $(CFLAGS_AOS4) $(INCLUDES) -o $@ src/iGame_cat.c
+src/iGame_cat_OS4.o: src/iGame_cat.c
+	$(CC_PPC) $(CFLAGS_OS4) $(INCLUDES_OS4) -o $@ src/iGame_cat.c
 
 ##########################################################################
 # generic build options
@@ -214,30 +224,21 @@ clean:
 
 # pack everything in a nice lha file
 release:
-	rm -rf iGame_rel/iGame/
-	rm iGame_rel/iGame.info
-	mkdir iGame_rel/iGame-$(DATE)
-	echo "#ifndef VERSION" > src/version.h
-	echo "#define VERSION \"\$$VER:version 2.0b4 ("$(DATE)")"\" >> src/version.h
-	echo "#endif" >> src/version.h
-	make iGame -f Makefile.linux
-	make iGame.030 -f Makefile.linux
-	make iGame.040 -f Makefile.linux
-	make iGame.060 -f Makefile.linux
-	make iGame.MOS -f Makefile.linux
-	cp iGame iGame_rel/iGame-$(DATE)/
-	cp iGame.030 iGame_rel/iGame-$(DATE)/
-	cp iGame.040 iGame_rel/iGame-$(DATE)/
-	cp iGame.060 iGame_rel/iGame-$(DATE)/
-	cp iGame.MOS iGame_rel/iGame-$(DATE)/
-	cp required_files/genres iGame_rel/iGame-$(DATE)/
-	cp required_files/igame.iff iGame_rel/iGame-$(DATE)/
-	cp required_files/iGame.info iGame_rel/iGame-$(DATE)/
-	cp required_files/igame_drawer_3.0.info iGame_rel/iGame.info
-	cp guide/* iGame_rel/iGame-$(DATE)/
+	cp required_files iGame_rel/iGame-$(DATE) -r
+	cp alt_icons iGame_rel/iGame-$(DATE)/Icons -r
+	cp iGame_rel/iGame-$(DATE)/igame_drawer_3.0.info iGame_rel/iGame-$(DATE).info
+	mv iGame_rel/iGame-$(DATE)/igame_drawer_3.0.info iGame_rel/iGame-$(DATE)/Icons/
+	mv iGame_rel/iGame-$(DATE)/igame_drawer.info iGame_rel/iGame-$(DATE)/Icons/
+	if [ -f "iGame" ]; then cp iGame iGame_rel/iGame-$(DATE)/; fi
+	if [ -f "iGame.030" ]; then cp iGame.030 iGame_rel/iGame-$(DATE)/; fi
+	if [ -f "iGame.040" ]; then cp iGame.040 iGame_rel/iGame-$(DATE)/; fi
+	if [ -f "iGame.060" ]; then cp iGame.060 iGame_rel/iGame-$(DATE)/; fi
+	if [ -f "iGame.MOS" ]; then cp iGame.MOS iGame_rel/iGame-$(DATE)/; fi
+	if [ -f "iGame.OS4" ]; then cp iGame.OS4 iGame_rel/iGame-$(DATE)/; fi
 	cp CHANGELOG.md iGame_rel/iGame=$(DATE)/
-	mkdir iGame_rel/iGame-$(DATE)/Icons/
-	cp alt_icons/* iGame_rel/iGame-$(DATE)/Icons/
-	cp required_files/igame_drawer.info iGame_rel/iGame-$(DATE)/Icons.info
-	cd iGame_rel && mv iGame-$(DATE) iGame && lha -a iGame-$(DATE).lha iGame/ && lha -a iGame-$(DATE).lha iGame.info
+	cd iGame_rel && lha -aq2o6 iGame-$(DATE).lha iGame-$(DATE)/ iGame-$(DATE).info
 
+clean-release:
+	rm -rf iGame_rel/iGame-$(DATE)
+	rm iGame_rel/iGame-$(DATE).lha
+	rm iGame_rel/iGame-$(DATE).info
